chore: fix test and refactor
